:fire: fusionne les deux fichiers entry.prod et entry.dev dans un seul fichier
On veut éviter de se répeter (DRY)
